requirements.txt
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
OWSLib==0.25.0
pyproj==3.2.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
psycopg2-binary==2.9.3
psycopg2==2.9.3
pyshp==2.3.1
xmltodict==0.13.0
shapely==2.0.1
beautifulsoup4==4.11.2
lxml==4.9.2
python-dotenv==1.0.0 





